fix the problem that made the program not working with the new txt file

+++ rental Project/customers/CommercialCustomers.cs
@@ -7,7 +7,7 @@
             int numberOfMonths,
             string theCarModel,
             int theCarModelYear,
-            int theBasePrice,
+            double theBasePrice,
             IMember memberType)
         {
             this.ID = ID;
